requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.92
argparse==1.3.0
elasticsearch==1.3.0
elasticsearch-dsl==0.0.2
facepy==1.0.4
itsdangerous==0.24
pbr==0.10.7
python-dateutil==2.3
pytz==2014.10
requests==2.5.1
rethinkdb==1.16.0-2
six==1.9.0
stevedore==1.2.0
urllib3==1.10
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
wsgiref==0.1.2
